Company listing,deleteing and updating view create
In this Commit if admin logged in then all the companies will be
returned to him so he can access any company profile and can edit their
profile and can delete any company.
In the commit also little changes to some other views are also made to
clarify code or to impose some rules on them.