CRUD for Employee with Filters Completed and also AddTeamView updated
This Commit contain Crud Operation(Create, Update, List, Delete) of Employee. Also  handling permissions of user to access company or if they are allow to perform crud in that company for Employee. Commit Also contain Filters and Pagination classes attached with CRUD Views.
Also A previous AddteamView is also updated as it now include a method
instead of some lines of code as that code was getting used in many
other views so impprove duplication a generic method added.