MiddleWare created, Ordering of user and company added, users search updated
This Commit majorly contain MiddleWare to make sure the security for
each company so only owner of that company or manager who have access of
that company can access the comapny and can play with that, previously
this was done but each view has same duplication to achieve the same
security functionality but Now middleware and a custom permision
handling this Scenrio in a smooth way without code duplications.
Also by this we have achieved No duplication at most of the places even
middleware is not placed on all views only on auth, Comapny team and
Employee views.
Also Users search field updated and ordering added for listing of users
and for companies as well.