Delete endpoint of Team, Employee, Payroll updated.
Team delete enpoint changed to a list of teams which can be deleted on
one click.
Employee delete endpooint changed to a list of employees_list so on one
click number of employees can be deleted.
Payroll delete view  now have same functionality as in team and payroll.
Also a new endpoint created for Retrieving data for a single  employee.